requirements.txt
# ============================================================================
# CONSOLIDATED ML PLATFORM - COMPLETE REQUIREMENTS
# FastAPI + Kedro ML Engine + EDA Integration
# Python 3.8+ (Tested with 3.9-3.12)
# ============================================================================

# ============================================================================
# CORE WEB FRAMEWORK
# ============================================================================
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# ============================================================================
# DATABASE & ORM
# ============================================================================
sqlalchemy>=2.0.0
PyMySQL>=1.1.0
mysql-connector-python>=8.0.0
alembic>=1.12.0

# ============================================================================
# SECURITY & AUTHENTICATION
# ============================================================================
cryptography>=40.0.0
PyJWT>=2.6.0
passlib[bcrypt]>=1.7.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# ============================================================================
# KEDRO ML FRAMEWORK (1.1.1 - Latest)
# ============================================================================
kedro==1.1.1
kedro-datasets[pandas]==1.5.0

# ============================================================================
# DATA PROCESSING & NUMERICAL
# ============================================================================
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
xgboost>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.2

# ============================================================================
# DATA ANALYSIS & VISUALIZATION
# ============================================================================
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
Pillow>=10.0.0

# ============================================================================
# ASYNC & TASK QUEUE
# ============================================================================
celery>=5.3.0
redis>=5.0.0

# ============================================================================
# HTTP & API
# ============================================================================
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.9.0

# ============================================================================
# CONFIGURATION & UTILITIES
# ============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
python-json-logger>=2.0.0
structlog>=23.0.0

# ============================================================================
# TYPE CHECKING & CODE QUALITY
# ============================================================================
typing-extensions>=4.9.0
mypy>=1.8.0

# ============================================================================
# TESTING (Development)
# ============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.13.0
faker>=22.0.0

# ============================================================================
# CODE QUALITY (Development)
# ============================================================================
black>=24.1.0
flake8>=7.0.0
pylint>=3.0.0
isort>=5.13.0
bandit>=1.7.0

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================
ipython>=8.0.0
jupyter>=1.0.0
notebook>=7.0.0
rich>=13.0.0
